[Refactor]: Auth 모듈 삭제, User 모듈 Public Controller로 구분, Feign Client로 요청 시 권한으로 인한 에러 발생 문제 해결

+++ common/src/main/java/com/codeterian/common/infrastructure/util/CustomAuditorAware.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 
 @Component
 @Slf4j
-public class CustomAuditorAware implements AuditorAware<Integer> {
+public class CustomAuditorAware implements AuditorAware<Long> {
 
 
     private final TokenUtils tokenUtils;
@@ -26,7 +26,7 @@ public class CustomAuditorAware implements AuditorAware<Integer> {
 
 
     @Override
-    public Optional<Integer> getCurrentAuditor() {
+    public Optional<Long> getCurrentAuditor() {
 
         HttpServletRequest request = ((ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()).getRequest();
